xtask: Refactor dist command
